Executing process app inside docker
