+++ requirements.txt
@@ -1,8 +1,18 @@
 attrs==17.3.0
+certifi==2017.11.5
+chardet==3.0.4
+docker==2.7.0
+docker-pycreds==0.2.1
 falcon==1.3.0
 gunicorn==19.7.1
+idna==2.6
+mock==2.0.0
+pbr==3.1.1
 pluggy==0.6.0
 py==1.5.2
 pytest==3.3.1
 python-mimeparse==1.6.0
+requests==2.18.4
 six==1.11.0
+urllib3==1.22
+websocket-client==0.46.0
