+++ runner/tests/Dockerfile_dummy
@@ -1,0 +1,3 @@
+FROM debian:jessie-slim
+
+ENTRYPOINT ["echo", "hello"]
