+++ settings.py
@@ -1,0 +1,6 @@
+from runner import processors
+
+
+DEBUG = False
+
+PROCESSOR = processors.DockerProcessor()
